fix: 프로필 사진 변경 안되는 오류 수정

+++ src/main/java/com/example/goguma/service/PwService.java
@@ -34,14 +34,32 @@ public class PwService {
         User user = userRepository.findById(id).orElseThrow(
                 NoSuchUserException::new
         );
-        if (user.getProfilePic() != null){
-            String[] spliturl = user.getProfilePic().split("https://gogumacat-s3.s3.ap-northeast-2.amazonaws.com/");
-            s3Service.delete(spliturl[1]);
-        }
-        String name = s3Service.uploadToAWS(profileUpdateDto.getProfilePic());
-        String profilePic = "https://gogumacat-s3.s3.ap-northeast-2.amazonaws.com/" + name;
-        String password = passwordEncoder.encode(profileUpdateDto.getPassword());
+        if (profileUpdateDto.getProfilePic() != null){
+            if(user.getProfilePic().contains("kakaocdn")){
 
-        user.update(profileUpdateDto, password,profilePic);
+                String name = s3Service.uploadToAWS(profileUpdateDto.getProfilePic());
+                String profilePic = "https://gogumacat-s3.s3.ap-northeast-2.amazonaws.com/" + name;
+                String password = passwordEncoder.encode(profileUpdateDto.getPassword());
+                user.update(profileUpdateDto, password,profilePic);
+            }
+
+            if(user.getProfilePic() != null){
+                String[] spliturl = user.getProfilePic().split("https://gogumacat-s3.s3.ap-northeast-2.amazonaws.com/");
+                s3Service.delete(spliturl[1]);
+
+                String name = s3Service.uploadToAWS(profileUpdateDto.getProfilePic());
+                String profilePic = "https://gogumacat-s3.s3.ap-northeast-2.amazonaws.com/" + name;
+                String password = passwordEncoder.encode(profileUpdateDto.getPassword());
+                user.update(profileUpdateDto, password,profilePic);
+
+            }
+        } else {
+            String profilePic = user.getProfilePic();
+            String password = passwordEncoder.encode(profileUpdateDto.getPassword());
+            user.update(profileUpdateDto, password,profilePic);
+        }
+
+
+
     }
 }
